// src/core/enums.ts
export enum CrewMemberId {
  NA = -1,
  Commander = 0,
  Alyssa = 1,
  Armatur = 2,
  Bentos = 3,
  C2U2 = 4,
  ChiTi = 5,
  Crystal = 6,
  Dane = 7,
  Deirdre = 8,
  Doc = 9,
  Draco = 10,
  Iranda = 11,
  Jeremiah = 12,
  Jujubal = 13,
  Krydon = 14,
  Luis = 15,
  Mercedez = 16,
  Milete = 17,
  MuriL = 18,
  Mystyc = 19,
  Nandi = 20,
  Orestes = 21,
  Pancho = 22,
  PS37 = 23,
  Quarck = 24,
  Sosumi = 25,
  Uma = 26,
  Wesley = 27,
  Wonton = 28,
  Yorvick = 29,
  Zeethibal = 30,
  Opponent = 31,
  Wild = 32,
  Jarek = 33,
  FamousCaptain = 34,
  Dragonfly = 35,
  Scarab = 36,
  SpaceMonster = 37,
  Aragorn = 38,
  Brady = 39,
  EightOfNine = 40,
  Fangorn = 41,
  Gagarin = 42,
  Hoshi = 43,
  Jackson = 44,
  Kaylee = 45,
  Marcus = 46,
  ONeill = 47,
  Ripley = 48,
  Stilgar = 49,
  Taggart = 50,
  Vansen = 51,
  Xizor = 52,
  Princess = 53,
  Scorpion = 54,
}

export enum Difficulty {
  Beginner = 0,
  Easy = 1,
  Normal = 2,
  Hard = 3,
  Impossible = 4,
}
// ...
export enum OpponentType {
  Bottle,
  FamousCaptain,
  Mantis,
  Pirate,
  Police,
  Trader,
}
// ...
export enum ShipType {
  Flea = 0,
  Gnat = 1,
  Firefly = 2,
  Mosquito = 3,
  Bumblebee = 4,
  Beetle = 5,
  Hornet = 6,
  Grasshopper = 7,
  Termite = 8,
  Wasp = 9,
  SpaceMonster = 10,
  Dragonfly = 11,
  Mantis = 12,
  Scarab = 13,
  Bottle = 14,
  Custom = 15,
  Scorpion = 16,
}
// ...
export enum StarSystemId {
  NA = -1,
  Acamar = 0,
  Adahn = 1,
  Aldea = 2,
  Andevian = 3,
  Antedi = 4,
  Balosnee = 5,
  Baratas = 6,
  Brax = 7,
  Bretel = 8,
  Calondia = 9,
  Campor = 10,
  Capelle = 11,
  Carzon = 12,
  Castor = 13,
  Cestus = 14,
  Cheron = 15,
  Courteney = 16,
  Daled = 17,
  Damast = 18,
  Davlos = 19,
  Deneb = 20,
  Deneva = 21,
  Devidia = 22,
  Draylon = 23,
  Drema = 24,
  Endor = 25,
  Esmee = 26,
  Exo = 27,
  Ferris = 28,
  Festen = 29,
  Fourmi = 30,
  Frolix = 31,
  Gemulon = 32,
  Guinifer = 33,
  Hades = 34,
  Hamlet = 35,
  Helena = 36,
  Hulst = 37,
  Iodine = 38,
  Iralius = 39,
  Janus = 40,
  Japori = 41,
  Jarada = 42,
  Jason = 43,
  Kaylon = 44,
  Khefka = 45,
  Kira = 46,
  Klaatu = 47,
  Klaestron = 48,
  Korma = 49,
  Kravat = 50,
  Krios = 51,
  Laertes = 52,
  Largo = 53,
  Lave = 54,
  Ligon = 55,
  Lowry = 56,
  Magrat = 57,
  Malcoria = 58,
  Melina = 59,
  Mentar = 60,
  Merik = 61,
  Mintaka = 62,
  Montor = 63,
  Mordan = 64,
  Myrthe = 65,
  Nelvana = 66,
  Nix = 67,
  Nyle = 68,
  Odet = 69,
  Og = 70,
  Omega = 71,
  Omphalos = 72,
  Orias = 73,
  Othello = 74,
  Parade = 75,
  Penthara = 76,
  Picard = 77,
  Pollux = 78,
  Quator = 79,
  Rakhar = 80,
  Ran = 81,
  Regulas = 82,
  Relva = 83,
  Rhymus = 84,
  Rochani = 85,
  Rubicum = 86,
  Rutia = 87,
  Sarpeidon = 88,
  Sefalla = 89,
  Seltrice = 90,
  Sigma = 91,
  Sol = 92,
  Somari = 93,
  Stakoron = 94,
  Styris = 95,
  Talani = 96,
  Tamus = 97,
  Tantalos = 98,
  Tanuga = 99,
  Tarchannen = 100,
  Terosa = 101,
  Thera = 102,
  Titan = 103,
  Torin = 104,
  Triacus = 105,
  Turkana = 106,
  Tyrus = 107,
  Umberlee = 108,
  Utopia = 109,
  Vadera = 110,
  Vagra = 111,
  Vandor = 112,
  Ventax = 113,
  Xenon = 114,
  Xerxes = 115,
  Yew = 116,
  Yojimbo = 117,
  Zalkon = 118,
  Zuul = 119,
  Centauri = 120,
  Galvon = 121,
  Inthara = 122,
  Meridian = 123,
  Qonos = 124,
  Rae = 125,
  Weytahn = 126,
  Zonama = 127,
}